nosql/atividade-by-vitao/requirements.txt
pymongo==4.6.0
folium==0.15.0
python-dotenv==1.0.0
